python_backend/requirements.txt
flask
crewai
crewai-tools
langchain_google_genai
python-dotenv
google-generativeai
